Squashed 'src/' changes from 2c5fcbaa..8fb64150
8fb64150 Merge branch 'master' into module
e37b8fc8 Issue 160 in phreeqc. llnl minimum temperature problem. There are several places where a SOLUTION is constructed to get lists or test the database. By default the SOLUTION is 25C, so if llnl has a minimum temperature greater (or maximum temperature less) than 25, an error is encountered. The fix is to change the temperature of the SOLUTIONs to the minimum temperature of the llnl grid if llnl is being used.

git-subtree-dir: src
git-subtree-split: 8fb64150bb19126d4f0782dcb8e718caa7ded80c